+++ api/src/api/lvc.hpp
@@ -106,8 +106,8 @@ extern "C" {
     } LvcCreateInput;
 
     typedef struct {
-        uint32_t line_start;
-        uint32_t line_end;
+        size_t line_start;
+        size_t line_end;
         const char* content;
     } LvcDiff;
 
@@ -118,7 +118,7 @@ extern "C" {
     } LvcConflict;
 
     typedef struct {
-        uint32_t length;
+        size_t length;
         LvcConflict* conflict;
     } LvcConflictArray;
 
@@ -148,7 +148,7 @@ extern "C" {
     LVC_API LvcError lvc_sync(const char* lvc, const char* tmp_lvc, LvcConflictArray* conflicts) noexcept;
     LVC_API LvcError lvc_unite(const char* lvc, const char* src_workspace_name, const char* dest_workspace_name, LvcConflictArray* conflicts) noexcept;
     LVC_API LvcError lvc_insert(const char* lvc, const char* src_workspace_name, const char* dest_workspace_name, LvcConflictArray* conflicts) noexcept;
-    LVC_API LvcError lvc_revert(const char* lvc, char* version_id, uint32_t input_count, char** input) noexcept;
+    LVC_API LvcError lvc_revert(const char* lvc, char* version_id, size_t input_count, char** input) noexcept;
     LVC_API LvcError lvc_push_local(const char* lvc, const char* tmp_lvc) noexcept;
     LVC_API LvcError lvc_push_server(const char* lvc) noexcept;
     LVC_API LvcError lvc_rename_category(const char* lvc, const char* category_name, const char* new_name) noexcept;
